see notes
Note that file names for controllers don't necessarily match the class names. This is harder to fix than to work with as far as I can tell. Unwind segues should work now, the map segue is proven to work and the rest are set up accordingly. Simply a back button press will send it to the previous screen! This allows us to pass data between them more easily. Everything is also now classed.