PLEASE READ
Changes are in the maps view controller class file. Things are a bit messy but you can cmd+f "requestURL" for where parsing and such begins. Most of it is fairly self-explanatory.
Additional info: I fixed location authorization and should be able to add pins soon. I believe this was the one step keeping me from customizing pins usefully before. This means that tomorrow I should be able to implement some custom pins that when you select the info button on them should send you to the page with the list of reviews for that location.