Add files via upload
Created a starting viewcontroller for the app with animation. I decided not to put all of this on the launch screen as that only persists until the app is fully loaded, and making it persist longer appears to be bad programming practice. I didn't wire it up to any of the other views, and I didn't make NapQuest one word but those are totes easy to fix. Also, I changed the rating icons from stars to empty circles and circles with "zzz" inside them